Fixed bug where some columns don't render when width of frozen columns exceeded width of scrollable area for normal columns.
Rearranged some variables so it is more apparent where there are two
similar variables (one for the frozen columns sections and one for
normal columns).

Fixed an error where frozen columns were being interpreted as being
outside the viewport and then causing the entire rendering loop to
"break;" and not render any subsequent frozen columns or any of the
non-frozen columns.