Use of navigator.userAgent instead of $.browser
Copied from merge of pull request 521 on mleibman master Jan 21, 2013

Note by dholcombe: This does not eliminate all usage of $.browser. There
is still a check for $.browser.msie. If that isn't fixed in another
mleibman commit by the time I catch up to current I need to go eliminate
that use and submit a pull request to mleibman.